requirements.txt
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.4
django-cors-headers==3.1.0
django-rest-swagger==2.2.0
django-safedelete==0.5.1
django-taggit==1.1.0
djangorestframework==3.10.2
django-test
future==0.17.1
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==6.1.0
psycopg2==2.8.3
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3

